rename chunithm_net._httpx to chunithm_net._auth
